requirements.txt
# 核心依赖
requests>=2.31.0
pyyaml>=6.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas>=2.0.3
tqdm>=4.66.1
python-dotenv>=1.0.0

# 爬虫引擎依赖
playwright>=1.41.0  # 浏览器自动化
firecrawl-sdk>=1.0.0  # Firecrawl SDK

# AI分析依赖
google-generativeai>=0.3.1  # Google Gemini
openai>=1.0.0  # OpenAI API

# 通知依赖
dingtalk-stream>=0.14.0  # 钉钉通知
feishu-stream>=0.1.0     # 飞书通知

# 工具依赖
click>=8.1.7  # 命令行参数处理
rich>=13.5.2  # 命令行格式化输出
retry>=0.9.2  # 重试机制
tenacity>=8.2.3  # 高级重试机制
pytz>=2023.3  # 时区处理
jsonschema>=4.19.0  # JSON验证
cryptography>=41.0.3  # 加密支持
pytest>=7.4.0  # 测试框架
pytest-playwright>=0.4.2  # Playwright测试支持

# 网络和HTML解析
html5lib>=1.1
httpx>=0.23.0
fake-useragent>=1.0.0

# 数据处理
numpy>=1.22.0
openpyxl>=3.0.0  # Excel支持
tabulate>=0.9.0  # 表格格式化

# 调试和开发工具
pytest>=7.0.0  # 测试

# 增强的爬虫工具
selenium>=4.9.0  # Selenium浏览器自动化
webdriver-manager>=3.8.6  # WebDriver管理
pydantic>=2.0.0  # 用于数据模型和验证 